# core-module/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Национальный сервис мониторинга цен на продовольственные товары(OpenAPI 3.0)
  description: |-
    Национальный сервис мониторинга цен на продовольственные товары
    
    Полезные ссылки:
    - [Репозиторий](https://github.com/AnnaDesert/Learning)
  version: 1.0.11
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /shop:
    get:
      tags:
        - shop
      summary: Получить все торговые точки
      description: Получить все торговые точки
      operationId: allShop
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
    post:
      tags:
        - shop
      summary: Добваить торговую точку
      description: Добваить торговую точку
      operationId: addShop
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopCreationDTO'
        required: true
      responses:
        '201':
          description: Created
  /shop/{id}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по ID
      description: Получить торговую точку по ID
      operationId: getShopById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ShopDTO'
    delete:
      tags:
        - shop
      summary: Удалить торговую точку по ID
      description: Удалить торговую точку по ID
      operationId: removeShopById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - shop
      summary: Изменить торговую точку
      description: Изменить торговую точку
      operationId: updateShop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopCreationDTO'
        required: true
      responses:
        '200':
          description: OK
  /shop/name/{name}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по названию
      description: Получить торговую точку по названию
      operationId: getShopByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
  /shop/address/{address}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по адресу
      description: Получить торговую точку по адресу
      operationId: getShopByAddress
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
...
  /category:
    get:
      tags:
        - category
      summary: Получить все категории товара
      description: Получить все категории товара
      operationId: allCategory
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductCategoryDTO'
    post:
      tags:
        - category
      summary: Добваить категорию товара
      description: Добваить категорию товара
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategoryCreationDTO'
        required: true
      responses:
        '201':
          description: Created
  /category/{id}:
    get:
      tags:
        - category
      summary: Получить категорию товара по ID
      description: Получить категорию товара по ID
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductCategoryDTO'
    delete:
      tags:
        - category
      summary: Удалить категорию товара по ID
      description: Удалить категорию товара по ID
      operationId: removeCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - category
      summary: Изменить категорию товара
      description: Изменить категорию товара
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategoryCreationDTO'
        required: true
      responses:
        '200':
          description: OK
  /category/name/{name}:
    get:
      tags:
        - category
      summary: Получить категорию товара по названию
      description: Получить категорию товара по названию
      operationId: getCategoryByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductCategoryDTO' 
...
components:
  schemas:
...
    ShopCreationDTO:
      required:
        - address
        - idAdmin
        - name
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        idAdmin:
          type: integer
          format: int64
...
    ProductCategoryCreationDTO:
      required:
        - name
      type: object
      properties:
        name:
          type: string
...
    ShopDTO:
      required:
        - address
        - id
        - idAdmin
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        idAdmin:
          type: integer
          format: int64
...
    ProductCategoryDTO:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...