# event-manager-module/target/generated-sources/swagger/main/java/org/senla/swagger/api/ShopApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.55).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package org.senla.swagger.api;

import org.senla.swagger.model.ShopCreationDTO;
import org.senla.swagger.model.ShopDTO;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import javax.validation.constraints.*;
import java.io.IOException;
import java.util.List;
import java.util.Map;
import java.util.Optional;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2024-06-24T09:46:36.596518200+03:00[Europe/Moscow]")
@Validated
public interface ShopApi {

    Logger log = LoggerFactory.getLogger(ShopApi.class);

    Optional<ObjectMapper> getObjectMapper();

    Optional<HttpServletRequest> getRequest();

    

    @Operation(summary = "Добваить торговую точку", description = "Добваить торговую точку", tags={ "shop" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "201", description = "Created") })
    @RequestMapping(value = "/shop",
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> addShop(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody ShopCreationDTO body
);


    @Operation(summary = "Получить все торговые точки", description = "Получить все торговые точки", tags={ "shop" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = ShopDTO.class)))) })
    @RequestMapping(value = "/shop",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<ShopDTO>> allShop();


    @Operation(summary = "Получить торговую точку по адресу", description = "Получить торговую точку по адресу", tags={ "shop" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = ShopDTO.class)))) })
    @RequestMapping(value = "/shop/address/{address}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<ShopDTO>> getShopByAddress(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("address") String address
);


    @Operation(summary = "Получить торговую точку по ID", description = "Получить торговую точку по ID", tags={ "shop" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "*/*", schema = @Schema(implementation = ShopDTO.class))) })
    @RequestMapping(value = "/shop/{id}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<ShopDTO> getShopById(@Min(1L)@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema(allowableValues={ "1" }, minimum="1"
)) @PathVariable("id") Long id
);


    @Operation(summary = "Получить торговую точку по названию", description = "Получить торговую точку по названию", tags={ "shop" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = ShopDTO.class)))) })
    @RequestMapping(value = "/shop/name/{name}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<ShopDTO>> getShopByName(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("name") String name
);


    @Operation(summary = "Удалить торговую точку по ID", description = "Удалить торговую точку по ID", tags={ "shop" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK") })
    @RequestMapping(value = "/shop/{id}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> removeShopById(@Min(1L)@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema(allowableValues={ "1" }, minimum="1"
)) @PathVariable("id") Long id
);


    @Operation(summary = "Изменить торговую точку", description = "Изменить торговую точку", tags={ "shop" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK") })
    @RequestMapping(value = "/shop/{id}",
        consumes = { "application/json" }, 
        method = RequestMethod.PATCH)
    ResponseEntity<Void> updateShop(@Min(1L)@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema(allowableValues={ "1" }, minimum="1"
)) @PathVariable("id") Long id
, @Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody ShopCreationDTO body
);

}